codesplit, fix edit dialog, add jump link
