+++ components/locales/de/component-return.json
