+++ components/locales/de/component-start.html
@@ -1,0 +1,48 @@
+<script type="text/html" data-help-name="component_in">
+    <p>Startknoten für den Flow einer Komponente.</p>
+    <h3>Inputs</h3>
+        <p>
+        Startknoten definieren ihre erwarteten Parameter über die API. 
+        Je Parameter kann der Name des message parts und sein Typ angegeben werden. 
+        Zusätzlich kann der Parameter auf "erwartet" gesetzt werden.
+        </p>
+
+        <h4>Option `global`</h4>
+        <ul>
+            <li>
+                <b>eingeschaltet</b>
+                Der Parameter wird direkt in die <code>msg</code> gesetzt und dort auch nach dem Verlassen des Flows der Komponente nicht wieder entfernt. 
+                Er bleibt also "global" in der <code>msg</code> erhalten.
+            </li>
+            <li>
+                <b>ausgeschaltet</b>
+                Der Parameter ist nur "lokal" verfügbar, also nur innerhalb des Flows der Komponente, in <code>msg.component</code>. Danach wird er wieder entfernt.
+                Ein lokaler Parameter wird durch die Komponente automatisch auf dem Stack gesichert, so dass er bei geschachtelten Aufrufen nicht verloren geht.
+            </li>
+        </ul>
+        Default ist "aus", i.e. "lokal"
+ </script>
+
+<script type="text/html" data-help-name="component">
+    <p>
+        Knoten, der eine ausgewählte Komponente ausführt. 
+        Im Editor kann aus allen definierten Komponenten eine ausgewählt werden, die von diesem Knoten aufgerufen werden soll.
+        Nach der Auswahl erscheinen alle Parameter der API der Komponente und es kann bestimmt werden, aus welcher Quelle der jeweilge
+        Parameter seinen Wert erhalten soll. Dies erspart das Vorbereiten von message parts durch Change Knoten etc.
+    </p>
+    <p>
+        Parameter, die erwartet werden, müssen mit einem gültigen Wert definiert werden (nicht <code>null</code> und nicht <code>undefined</code>),
+        sonst erscheint beim Ausführen der Komponente ein Fehler.
+    </p>
+    <h3>Inputs</h3>
+        <p>
+        Je Parameter wird hier ein Wert konfiguriert, der an die Komponente in der message übergeben wird.
+        </p>
+ </script>
+
+<script type="text/html" data-help-name="component_out">
+    <p>
+        Endknoten für den Flow einer Komponente.
+        Der Flow wird beendet und der aufrufende Knoten übernimmt die weitere Ausführung.
+    </p>
+ </script>
